Upgrade Harvest to no longer check for primary products
The concept of skipping secondary products is no longer relevant with the current use of the legacy harvest on EN machines. We will manage the data accordingly to avoid loading products that should not be loaded.

This was causing erroneous skips of products because harvest was tracking secondary members for all bundles/collections encountered in a single run, not accounting for whether or not they are even within the same directory.

resolves #174